Initial work on megamenu component
