+++ app/screens/Explore.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { connect } from 'react-redux'
 import styled from 'styled-components/native'
-import { Platform, TouchableHighlight, Animated } from 'react-native'
+import { Platform, Animated } from 'react-native'
 import MapboxGL from '@react-native-mapbox-gl/maps'
 import { AndroidBackHandler } from 'react-navigation-backhandler'
 import Config from 'react-native-config'
@@ -45,6 +45,7 @@ import {
 import { bboxToTiles } from '../utils/bbox'
 import Header from '../components/Header'
 import MapOverlay from '../components/MapOverlay'
+import MegaMenu from '../components/MegaMenu'
 import AddPointOverlay from '../components/AddPointOverlay'
 import LoadingOverlay from '../components/LoadingOverlay'
 import ZoomToEdit from '../components/ZoomToEdit'
@@ -102,56 +103,6 @@ const MainHeader = styled(Header)`
 const StyledMap = styled(MapboxGL.MapView)`
   flex: 1;
   width: 100%;
-`
-
-const MegaMenu = styled.View`
-  flex: 1;
-  position: absolute;
-  justify-content: center;
-  align-items: center;
-  bottom: 16;
-  right: 16;
-`
-
-const MenuButton = styled.TouchableHighlight`
-  border-radius: 100;
-  width: 56;
-  height: 56;
-  background-color: ${colors.primary};
-  justify-content: center;
-  align-items: center;
-  elevation: 2;
-  shadow-color: ${colors.base};
-  shadow-opacity: 0.7;
-  shadow-offset: 0px 0px;
-`
-const AnimatedTouchable = Animated.createAnimatedComponent(TouchableHighlight)
-
-const ModeButton = styled(AnimatedTouchable)`
-  position: absolute;
-  border-radius: 100;
-  width: 48;
-  height: 48;
-  margin-bottom: 8;
-  background-color: ${colors.primary};
-  justify-content: center;
-  align-items: center;
-  shadow-color: ${colors.baseAlpha};
-  shadow-opacity: 0.7;
-  shadow-offset: 0px 0px;
-  elevation: 2;
-`
-
-const Label = styled(Animated.Text)`
-  color: white;
-  position: absolute;
-  right: 0;
-  text-align: right;
-  font-size: 14;
-  width: 100;
-  padding: 2px 4px;
-  border-radius: 4;
-  background-color: ${colors.baseMed};
 `
 
 class Explore extends React.Component {
@@ -451,86 +402,6 @@ class Explore extends React.Component {
 
     // if not in explicit mode, render default MapOverlay
 
-    const rotation = {
-      transform: [
-        {
-          rotate: this.animation.interpolate({
-            inputRange: [0, 1],
-            outputRange: ['0deg', '45deg']
-          })
-        }
-      ]
-    }
-
-    const opacityInterpolation = this.animation.interpolate({
-      inputRange: [0, 0.75, 1],
-      outputRange: [0, 0, 1]
-    })
-
-    const PointButtonPosition = {
-      opacity: opacityInterpolation,
-      transform: [
-        { scale: this.animation },
-        {
-          translateY: this.animation.interpolate({
-            inputRange: [0, 1],
-            outputRange: [0, -60]
-          })
-        }
-      ]
-    }
-
-    const WayButtonPosition = {
-      opacity: opacityInterpolation,
-      transform: [
-        { scale: this.animation },
-        {
-          translateY: this.animation.interpolate({
-            inputRange: [0, 1],
-            outputRange: [0, -116]
-          })
-        }
-      ]
-    }
-
-    const RecordButtonPosition = {
-      opacity: opacityInterpolation,
-      transform: [
-        { scale: this.animation },
-        {
-          translateY: this.animation.interpolate({
-            inputRange: [0, 1],
-            outputRange: [0, -172]
-          })
-        }
-      ]
-    }
-
-    const CameraButtonPosition = {
-      opacity: opacityInterpolation,
-      transform: [
-        { scale: this.animation },
-        {
-          translateY: this.animation.interpolate({
-            inputRange: [0, 1],
-            outputRange: [0, -228]
-          })
-        }
-      ]
-    }
-
-    const LabelPosition = {
-      opacity: opacityInterpolation,
-      transform: [
-        {
-          translateX: this.animation.interpolate({
-            inputRange: [0, 1],
-            outputRange: [0, -52]
-          })
-        }
-      ]
-    }
-
     return (
       <>
         <MapOverlay
@@ -538,37 +409,14 @@ class Explore extends React.Component {
           onAddButtonPress={this.onAddButtonPress}
           navigation={navigation}
         />
-        <MegaMenu>
-          <ModeButton style={[CameraButtonPosition]} underlayColor={colors.base} onPress={() => navigation.navigate('CameraScreen', { previousScreen: 'Explore', feature: null })}>
-            <>
-              <Label style={[LabelPosition]}>Take Photo</Label>
-              <Icon name='camera' size={20} color='#FFFFFF' />
-            </>
-          </ModeButton>
-          <ModeButton style={[RecordButtonPosition]} underlayColor={colors.base} status={currentTraceStatus} onPress={() => this.onRecordPress()}>
-            <>
-              <Label style={[LabelPosition]}>Record Track</Label>
-              <Icon name='circle-play' size={20} color='#FFFFFF' />
-            </>
-          </ModeButton>
-          <ModeButton style={[WayButtonPosition]} underlayColor={colors.base} onPress={() => { this.props.setMapMode(modes.ADD_WAY) }}>
-            <>
-              <Label style={[LabelPosition]}>Add Way</Label>
-              <Icon name='share-2' size={20} color='#FFFFFF' />
-            </>
-          </ModeButton>
-          <ModeButton style={[PointButtonPosition]} underlayColor={colors.base} onPress={() => { this.onAddButtonPress() }}>
-            <>
-              <Label style={[LabelPosition]}>Add Point</Label>
-              <Icon name='marker' size={20} color='#FFFFFF' />
-            </>
-          </ModeButton>
-          <MenuButton underlayColor={colors.base} onPress={this.toggleMenuOpen}>
-            <Animated.View style={[rotation]}>
-              <Icon name='plus' size={20} color='#FFFFFF' />
-            </Animated.View>
-          </MenuButton>
-        </MegaMenu>
+        <MegaMenu
+          onCameraPress={() => navigation.navigate('CameraScreen', { previousScreen: 'Explore', feature: null })}
+          onRecordPress={() => this.onRecordPress()}
+          onWayPress={() => { this.props.setMapMode(modes.ADD_WAY) }}
+          onPointPress={() => { this.onAddButtonPress() }}
+          onMenuPress={this.toggleMenuOpen}
+          recordStatus={currentTraceStatus}
+        />
       </>
     )
   }
